docs: Add typedoc to generate markdown for JS API
