+++ src/index.js
@@ -1,2 +1,3 @@
+export * from './prng-type.js';
 export * from './seeds.js';
 export * from './random-generator.js';
