+++ util/clean.mjs
@@ -4,7 +4,7 @@ import { fileURLToPath } from 'url';
     
 const __dirname = dirname(fileURLToPath(import.meta.url));
 
-['dist', 'bin'].forEach(dir => {
+['dist', 'bin', 'docs'].forEach(dir => {
     const dirPath = join(__dirname, '../', dir);
     rmSync(dirPath, { recursive: true, force: true });
 });
